requirements.txt
requests
Pillow
apscheduler
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dateutil
waveshare-epaper
cairosvg
filelock
astral
